dev-requirements.txt
pytest
pytest-cov
boto3
botocore
black
flake8
isort
pre-commit
bandit
pip-audit
awscli-local
requests
responses
jsonschema
jsonpath-ng
pyyaml
